Move ColonSplitter to core support package.
