+++ TelepresenceBot/core/build.gradle
@@ -6,3 +6,6 @@ dependencies {
     testCompile 'org.mockito:mockito-core:1.10.19'
     testCompile 'org.mockito:mockito-all:1.10.19'
 }
+
+sourceCompatibility = 1.7
+targetCompatibility = 1.7
