+++ TelepresenceBot/mobile/src/test/java/com/novoda/tpbot/automation/AutomationCheckerTest.java
@@ -1,5 +1,7 @@
 package com.novoda.tpbot.automation;
 
+import com.novoda.support.ColonStringSplitter;
+
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
